Update example to use Address format
It seems strange to me that the example for the From headers uses the full Address format, whereas the sender envelope does not.

I believe it is possible to use it here, so thought I would expand on the example to show how you can set both the display name and email address when configuring the sender envelope globally.