Changing `enum:` example away from APP_ENV
Reason: When I first read this, I was expecting that the Symfony-internal ways to get the environment (`{{ app.environment }}`, `$routingConfigurator->env()` ,etc.) would now return the enum. But this is not the case. So if you explain the new processor with exactly this example, there needs to be a caution box, explaining that Symfony's methods still return the string.

* If you don't agree with this change, I'd still open a separate PR to add this info (taken from https://symfony.com/blog/new-in-symfony-6-2-improved-enum-support#enums-in-environment-variables), cause that's a missing piece of information:

> The value stored in the ``CARD_SUIT`` env var would be a string like `'spades'` but the
> application will use the ``Suit::Spdes`` enum value.

* Besides, as the list of processors is getting longer and longer, I'd suggest to create a sub-heading for each (under "Built-In Environment Variable Processors"), to get rid of the endless indentation, and improve scanability, and get them included in the TOC.
Should I create a PR for this?