Switching to PHP-version-agnostic socket file
Page: https://symfony.com/doc/5.4/setup/web_server_configuration.html

This makes upgrading to a new PHP version so much easier (I just did that). If the PHP version is hardcoded in every domain's config file, it's a hassle to update them.

On the other hand: `/var/run/php/php-fpm.sock` is a symlink to the most recently installed PHP version, so this is what most people want probably.

I hope this doesn't interfere with the change I did in https://github.com/symfony/symfony-docs/pull/17882 on the 6.2 branch.