Delete Inky Email Templating Language documentation
The Inky framework has not been actively developed for about three years (last release in March 2022). In addition to Sass deprecations and other bugs, the project is generally no longer up to date when it comes to implementing email templates.

In my opinion, it no longer makes sense to list the project as the only solution for designed emails.